Product delete, CRUD of multiimage is added

+++ app/Http/Controllers/Backend/BrandController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Brand;
 use Image;
+use File;
 
 class BrandController extends Controller
 {   
@@ -73,25 +74,31 @@ class BrandController extends Controller
              'brand_name.required'=>'Please input Brand Name',
              'brand_image.mimes'=>'File format not supported',
         ]);
-        
-        $old_image = $request->old_image;
-
-        $brand_image = $request->file('brand_image');
-
-        if($brand_image){
-        $name_generate = hexdec(uniqid());
-        $image_ext = strtolower($brand_image->getClientOriginalExtension());
-        $image_name = 'img-'.$name_generate.'.'.$image_ext;
-        $image_location = 'backend/uploads/brand/';
-        $final_image = $image_location.$image_name;
-        Image::make($brand_image)->resize(300,300)->save($final_image);
-
-        unlink($old_image);
 
         $brand = Brand::findorFail($id);
+
+        $old_image = $request->old_image;
+        $brand_image = $request->file('brand_image');
+
+        if($brand_image)
+        {
+            $name_generate = hexdec(uniqid());
+            $image_ext = strtolower($brand_image->getClientOriginalExtension());
+            $image_name = 'img-'.$name_generate.'.'.$image_ext;
+            $image_location = 'backend/uploads/brand/';
+            $final_image = $image_location.$image_name;
+            Image::make($brand_image)->resize(300,300)->save($final_image);
+
+            if(File::exists($old_image))
+            {
+                unlink($old_image);
+            }
+
+            $brand->brand_image = $final_image;
+        }
+
         $brand->brand_name = $request->brand_name;
         $brand->brand_slug = strtolower(str_replace(' ', '-', $request->brand_name));
-        $brand->brand_image = $final_image;
         $brand->save();
 
         $notification = [
@@ -100,21 +107,6 @@ class BrandController extends Controller
         ];
 
         return redirect()->route('all.brands')->with($notification);
-
-        }
-        else{
-            $brand = Brand::findorFail($id);
-            $brand->brand_name = $request->brand_name;
-            $brand->brand_slug = strtolower(str_replace(' ', '-', $request->brand_name));
-            $brand->save();
-
-            $notification = [
-                'message' => 'Brand updated successfully.',
-                'alert-type' => 'info',
-            ];
-    
-            return redirect()->route('all.brands')->with($notification);
-        }
 
     }
